Implement ball/object collision
The intention here is to model objects as collections of surfaces. A
surface is essentially a line with an implied normal. If the ball gets
below a surface it gets pushed back out. Note that right now there is no
limit on how far an object will be pushed from.

Adjust the simulation to have two surfaces that make a V, randomize the
V shape and ball placement