Create first chamber
Replace the previous collision objects with a chamber plugin.

As of right now, the chamber plugin is compiled both as a wasm module to
be loaded in the browser, and compiled directly into the server. The
server side handles platform movement and physics, the wasm module is
only for rendering.

The plugin exposes a save/load mechanism. We save the state of the
plugin on each frame from the server, send that to the client, and load
the state back for rendering.

Resource embedding had to be updated to support pulling resources from
multiple locations. In our build we compile chamber.wasm and embed it in
the binary along with the rest of the resources from src/res

Collision handling was moved out of the main simulation and into a new
file physics.zig. For now this is where all our common code between
plugins and main live. Linear algebra types, ball definition, collision
handling etc. The vision here is that this will eventually evolve into a
libphysics that can be compiled into each wasm module.

Some of the collision handling was also simplified. The previous
intersectionPoint function has been adjusted to find the movement that
has to be done to undo the collision. This means we can just add the
movement to ball.pos. We also now only return that adjustment if a
collision actually occurred, this heavily simplifies collision handling
from the callers side.

Collision checking got some fixes as well. Before any time a ball was
below a platform it would get teleported up above the object. We forgot
that we should be checking that the ball actually traveled over the
surface

The simulation has been adjusted to run forever, but when the ball goes
out one side, it comes back in the other. This is just nicer in the
environment that our first chamber provides