Embed resources in application
This fixes the hardcoded cwd relative src/res http root and embeds
server resources directly in the application.

To keep the old development workflow, a --www-root flag has been added
which will be searched before looking at the embedded files.

Since we added an argument parser, we might as well add the ability to
specify a port to run the server on as well