Recording system
When things go wrong, it's nice to be able to reproduce them so we can
try out different fixes. Write down the current simulation state every
once in a while and expose a save button to flush it to disk.

The existing simulation was not deterministic. Specifically the floating
point precision errors could depend on how long our 1.6ms sleeps ended
up actually being. Instead step by a constant value until we hit the
current time.