Add multiple balls
Replace ball with balls, and handle collision between them. For initial
testing our chamber has been adjusted to have walls and a floor. With
the current physics systems when two balls collide on a surface, one can
push the other through the surface. This is a known bug and we are
choosing to live with it for now

Otherwise everything is pretty straightforward. The math for elastic
collision can be viewed at
https://en.wikipedia.org/wiki/Elastic_collision and has been simplified
with the assumption that all balls are the same mass