Run chamber.wasm on the server
The whole purpose of this project is to allow users to upload chamber
definitions and for us to run those chambers as part of our physics
simulation.

This requires some pretty heavy sandboxing. We don't want to just run
any arbitrary code, we need to ensure that the code can't do anything it
isn't supposed to do. Just linking an arbitrary shared object is not
sufficient here.

WASM is a good fit. It's reasonably efficient, and can be run both on
our server, can be used from any language, and in the web browser for
rendering. Users will only have to upload a single file, which we will
just run on the server and on the client.

Pull in wasmtime as our wasm engine. Engine choice was pretty poorly
researched. Hand wavey, it seemed like it had good backing, wasn't
overly complex, and had examples to help us get up and running quickly

* chamber.zig is guaranteed to be wasm now, so we can lose anything
  switching on wasm/not wasm
* Some chamber.zig interfaces had to be slightly adjusted to be exported
  with C ABI
* Sending balls in/out of wasm is a little sketchy. There isn't a strong
  guarantee that the struct layout on different platforms will be the
  same. For now we just assert that all fields are where we think they
  will be, and that happens to line up on both our x64 host, and in the
  wasm environment
* Due to some interface changes, SimulationHistory now owns a bunch of
  heap allocated data. In order to free it all, we need to walk from
  head->tail. This resulted in an iterator being pulled out as that
  iteration is now done in multiple places
* I don't have much to say about the actual wasm implementation. Use
  wasmtime, follow their examples.

  These were both helpful
  https://docs.wasmtime.dev/examples-c-hello-world.html
  https://docs.wasmtime.dev/examples-c-memory.html