Simplify chamber allocation interface
On alloc(), return the actual pointer to the data we want to write.
Previously we were returning a double pointer ?*[]u8, this was because a
slice is a complex structure which gets converted into a wasm out param.

This was fine, but complicates actual usage when passing data between
wasm and the host execution environment. It also forces implementers of
this interface to define a slicePtr and sliceLen interface, which isn't
necessarily reasonable.

We can get away with returning just a ?[*]u8, or a normal pointer, if we
just stash the length 4 bytes before what we return. This essentially
makes the VLA

```
struct {
  size_t len;
  char data[0];
}
```

but with manual packing of the len field;

Now callers can view this as a normal alloc where the returned pointer
can be written to directly