Prepare for multiple http servers
In test applications, we might want to spin up a quick http server to
demo some in progress functionality. The way the current http server was
set up it was not trivial to make an http server that served anything
but our existing ball state. Generalize a few of the existing concrete
types to make spinning up a new server simpler

* Extract a TcpServer that is an async tcp listener. On new connections
  it calls a new Spawner.spawn method which generates a new EventHandler
  to put on the event loop
* Move the HttpConnection concept into the http module with a generic
  HttpResponseGenerator input. This takes an http reader and returns an
  http writer
* The old Server now is a type that exports a Spawner for the TCP
  server. That spawner generates http connections with a
  HttpResponseGenerator that is tied to the server as well