Prepare for multiple http servers
In test applications, we might want to spin up a quick http server to
demo some in progress functionality. The way the current http server was
set up it was not trivial to make an http server that served anything
but our existing ball state. Generalize a few of the existing concrete
types to make spinning up a new server simpler

* Extract a TcpServer that is an async tcp listener. On new connections
  it calls a new Spawner.spawn method which generates a new EventHandler
  to put on the event loop
* Move the HttpConnection concept into the http module with a new "wait"
  state. In this state it tells whoever owns it that it needs to be fed
  with a response
* Wrap the HttpConnection with an object that polls it, and feeds it
  with the appropriate response