Improved chamber api
Chambers previously were forced to have allocation and deallocation
APIs, which resulted in chambers having to implement an allocator. In
zig this comes for free, but in other languages (C wasm-freestanding) we
are not so fortunate.

In reality, we do not actually need these APIs. What we actually need is
a segment of memory to write balls to, a section of memory to read the
canvas from, and a section of memory to read/write the save from. These
can be static memory if we want them to be.

Replace the existing alloc/free apis with an initialization param to set
static memory sizes etc. Externally we now call ballMemory(),
canvasMemory(), saveMemory() to figure out how to interact with the
module

While we're here, the concept of a state pointer no longer makes sense.
We are only going to initialize these chambers once per loaded module.
This is a great candidate for using global memory. Remove the State*
object