Add plinko chamber written in C
Export a C wrapper around physics, and use it to create a chamber in C

C wrapper intentionally duplicates the zig types. We want to allow the C
complier to inline what it needs to, and the ergonomics of working with
C structs are way better than if we were to use opaque pointers to zig
structs. We use some fancy assertions to ensure that the functions we
export in the C api match the zig types exactly

The plinko implementation revealed some issues with the physics system
* The elasticity calculations felt really broken. I haven't actually
  checked if the old math happens to line up with the new math, but my
  memory says that this change seemed to fix some extreme bounciness. If
  I were working at a real company, writing real code, I would force
  myself to double check. I don't, so we'll never know
* Add elasticity as a parameter to the collision application logic, the
  plinko pegs feel like they should absorb way more of the ball speed
  than the platforms in our other chambers do