Remove dependency on wasm_chamber from Simulation
In a future commit, we want to be able to run our simulation both on the
client, and on the server. Right now our simulation code depends
directly on our wasm chambers, which depend directly on wasmtime.

This will not compile to wasm

Replace the wasm_chamber.WasmChamber member in Simulation with an
interface that happens to be implemented by wasm_chamber.WasmChamber. A
future commit will instantiate the simulation with a different interface

Implementation notes
* Chambers now need to be stored as pointers to have stable memory
  addresses. The Chamber interface stores the WasmChamber pointer, which
  will be problematic if the chamber is on the stack and goes away.