Split out some js modules
This is prep work for enabling client side simulations. Move code
agnostic to where the simulation data is coming from into it's own
simulation_renderer module. Wrap this module in index.js with a
RemoteSimulation which fetches the data from the server and loads save
states before calling render

While we are here, we've fixed the wasmLog function not being
implemented.