Add a client side chamber testing page
After eefc6af49f6f1c5f9b90a6991bd3b9790878b89e, working on chambers kind
of sucks. Since chambers have to be added to the DB in order to be
displayed, we have to compile and upload a module every time we want to
see it. If we are iterating a lot, we end up with tons of duplicates in
the DB as we have no way to delete chambers right now

Add a client side testing page. This solves the previously mentioned
problem as well as allows people to work on chambers without ever
pulling the source code of this repo

Implementation details:
* Compile Simulation.zig to wasm and embed at /simulation.wasm
* Wrap Simulation.zig with wasm API that expects the browser to provide
  a chamber interface
* Load a chamber and simulation in the chamber test page. Forward
  Chamber api calls from simulation module to chamber module