Fix balls getting pushed through surfaces
If a collision pushed a ball through a surface, we would not detect it
as a collision. Our collision handling checks if a ball has moved
through a surface this frame in order to apply an elastic collision. If
we were more lax with this check we could end up applying the collision
more than once.

Add a second check to push an object out of a surface if it's detected
that it is there when it shouldn't be. This will only trigger if the
elastic collision detection is not hit

Limit the push to avoid yeeting balls too hard. We only need a small
impulse applied to prevent the issue in most cases

The right wall of the platforms chamber was also set incorrectly, which
resulted in pretty clear broken behavior after this fix