Define a real chamber height
Previously no chamber height was defined, but the client happened to
show a 600x450 chamber. This will be problematic in the future when
chambers are supposed to be above or below eachother in some way.
Hardcode a 0.7 height chamber in the server and expose that hardcoded
value to the client