Make all balls red
The red/blue differentiation starts to break down in a future commit
when balls can move between chambers. This was only introduced as a way
to visually distinguish between balls when testing physics.

Make all balls red, and add a stroked line around the edge to give a
little clearer definition