Use chamber DB ids directly
This will be helpful in a future commit when all chambers we may want to
look at will not be in the simulation. E.g. if a user wants to view the
modules that they have submitted and have not been accepted

This is yet another thing that we need to give to the client on start,
replace all individual getters with one big /init_info endpoint. We
could have left the other endpoints around, but that's just maintaining
dead code for the sake of API beauty. Just delete anything that we don't
need.

Linter script was failing due to another wasmtime jump on uninitialized
value.