Spinny bar chamber
Add a chamber that is just a spinning bar

This is the first chamber that has significant forces applied by moving
surfaces. Some slight adjustments had to be made to the physics library
to account for this. All collisions now take a object velocity that gets
added to the elastic velocity change

The spinning bar calculates the velocity of the surface at the location
of contact with the ball, and uses that to set the collision response.
This seems to work well enough