Add check build step
This creates a build step does not install the final executable. This
allows for the build system to not emit the binary, which is the most
expensive step by far. This step can be used by ZLS, which allows us to
turn on building on save without getting too slow, which gives us better
errors in the editor

We also skip resource embedding for this step, which reduces the cost of
rebuild on resource modification