Add chamber staging area
When a user submits a chamber, we should not just unconditionally add it
to the simulation. Add a staging state where chambers have to be
approved by an admin before becoming part of the main simulation

Staging state is defined by a new table in the DB accepted_chambers. If
the ID is in this list, it is part of the main simulation. APIs have
been adjusted to retrieve only accepted/unaccepted chambers

Add a new admin page, where we can preview the different chambers and
choose to accept or deny. This power is granted to a user who's ID
matches a new command line paramter --admin-id

Since we now have the concept of an admin, hide some APIs and UI
elements behind this admin user. Reject any requests to these endpoints
if the user is not the admin user

Add a second user to the testing DB, this allows us to test
authenticated and unauthenticated behavior in our integration scripts

With the new flow, it is a pain to set up a sample DB for testing,
update the generate_test_db program to embed all chambers immediately