Reduce simulation jitter under high latency
The way we were previously getting simulation state was pretty silly.
Every frame we were asking for the latest state, waiting for the
response, then rendering. If the latency gets above a few ms, it was
not possible to render at 60fps anymore

Update the /simulation_state API to return many states at once. This
allows us to buffer frames on the client and request more as needed

* Add a circular buffer impl to store history snapshots
* Save a snapshot every 10 physics steps
* Update /simulation_state to return all snapshots since the requested
  step number
* Update linter script