Add better failure messages to wasm failures
When a user uploads a module, we need to give some sort of feedback if
we fail the validation step. Add a way to attach detailed failure
messages to our errors. This will be more necessary in a future commit
when we start committing these failures to the DB. For now we just print
the error messages in test_chamber.

* This is modeled after std.*.Diagnostics (see std.tar, std.json, etc.)
* Traps and errors now populate some diagnostics, on error the
  diagnostics can be inspected for more useful info