DB support for validating chambers
If users are going to submit modules, we should have some form of
automated checks to ensure that they are ok to use. We need to store the
state of this validation in the database.

Replace the previous accepted_chambers table with a state field in the
chamber list. All chambers will now have a state associated with them

The new flow for chambers is to move from pending ->
validated/validation_failed -> accepted/rejected

In some cases validation may fail for a reason that we want to present
to a user, add an optional message to go with a chamber in the case of
failure. This is stored in a separate table

Bonus fixes:
* Delete unused deleteChamber